Reset port transciever if eeprom read fails.
Add a port transciever reset to the retry sequence that occurs
if there is a problem reading the transciever's eeprom.

The port reset involves putting the transciever into reset,
waiting one millisecond, taking the transciever out of
reset, and waiting for ten milliseconds.

Additionally, a ten millisecond wait was added after
initially taking the transciever out of reset.

The theory here is that some transcievers take more time
than others to come out of reset, so we add a short wait for
the transciever to set itself up properly.

Tags: chg, dev
TG-1729 # Closed

Change-Id: I906b47774cf6267f1d766f457b968b767459a438
Signed-off-by: Mike Berger <mabnhdev@gmail.com>